backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
google-adk
google-cloud-storage
psycopg2-binary
requests
pillow
pydantic